// web-application/src/types-openapi/api/models/AddOrderItemRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Badminton Court Management System API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface AddOrderItemRequest
 */
export interface AddOrderItemRequest {
    /**
     * 
     * @type {string}
     * @memberof AddOrderItemRequest
     */
    bookingId: string;
    /**
     * 
     * @type {number}
     * @memberof AddOrderItemRequest
     */
    productId: number;
    /**
     * 
     * @type {number}
     * @memberof AddOrderItemRequest
     */
    quantity: number;
}

/**
 * Check if a given object implements the AddOrderItemRequest interface.
 */
export function instanceOfAddOrderItemRequest(value: object): value is AddOrderItemRequest {
    if (!('bookingId' in value) || value['bookingId'] === undefined) return false;
    if (!('productId' in value) || value['productId'] === undefined) return false;
    if (!('quantity' in value) || value['quantity'] === undefined) return false;
    return true;
}

export function AddOrderItemRequestFromJSON(json: any): AddOrderItemRequest {
    return AddOrderItemRequestFromJSONTyped(json, false);
}

export function AddOrderItemRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): AddOrderItemRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'bookingId': json['bookingId'],
        'productId': json['productId'],
        'quantity': json['quantity'],
    };
}

export function AddOrderItemRequestToJSON(json: any): AddOrderItemRequest {
    return AddOrderItemRequestToJSONTyped(json, false);
}

export function AddOrderItemRequestToJSONTyped(value?: AddOrderItemRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'bookingId': value['bookingId'],
        'productId': value['productId'],
        'quantity': value['quantity'],
    };
}